requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
numpy
openai
tqdm
python-dotenv
langchain
langchain-openai
faiss-cpu
lxml
pdfplumber
requests
flask-cors
Flask
google-generativeai